move package info
